+++ backend/apid/graphql/entity.go
@@ -97,6 +97,26 @@ func (r *entityImpl) Related(p schema.EntityRelatedFieldResolverParams) (interfa
 		return []*types.Entity{}, err
 	}
 
+	// omit source
+	for i, en := range entities {
+		if en.ID != entity.ID {
+			continue
+		}
+
+		//
+		// - As we sort the result set in the next step we can safely remove the
+		//   source from the slice without preserving it's order.
+		// - Since we are dealing with a slice of pointers we explicilty set the
+		//   reference to the last element to nil to ensure it can be GC'd.
+		//
+		// https://github.com/golang/go/wiki/SliceTricks#delete-without-preserving-order
+		//
+		entities[i] = entities[len(entities)-1]
+		entities[len(entities)-1] = nil
+		entities = entities[:len(entities)-1]
+		break
+	}
+
 	// sort
 	scores := map[int]int{}
 	for i, en := range entities {
@@ -107,7 +127,7 @@ func (r *entityImpl) Related(p schema.EntityRelatedFieldResolverParams) (interfa
 		scores[i] = len(matched)
 	}
 	sort.Slice(entities, func(i, j int) bool {
-		return scores[i] < scores[j]
+		return scores[i] > scores[j]
 	})
 
 	// limit